approle tests part 2?

+++ src/VaultSharp/V1/AuthMethods/AppRole/Models/AppRoleRoleModel.cs
@@ -18,34 +18,34 @@ namespace VaultSharp.V1.AuthMethods.AppRole.Models
         public List<string> SecretIdBoundCIDRs { get; set; }
 
         [JsonProperty("secret_id_num_uses")]
-        public int SecretIdNumberOfUses { get; set; }
+        public long SecretIdNumberOfUses { get; set; }
 
         [JsonProperty("secret_id_ttl")]
-        public int SecretIdTimeToLive { get; set; }
+        public long SecretIdTimeToLive { get; set; }
 
         [JsonProperty("token_bound_cidrs")]
         public List<string> TokenBoundCIDRs { get; set; }
 
         [JsonProperty("token_explicit_max_ttl")]
-        public int TokenExplicitMaximumTimeToLive { get; set; }
+        public long TokenExplicitMaximumTimeToLive { get; set; }
 
         [JsonProperty("token_max_ttl")]
-        public int TokenMaximumTimeToLive { get; set; }
+        public long TokenMaximumTimeToLive { get; set; }
 
         [JsonProperty("token_no_default_policy")]
         public bool TokenNoDefaultPolicy { get; set; }
 
         [JsonProperty("token_num_uses")]
-        public int TokenNumberOfUses { get; set; }
+        public long TokenNumberOfUses { get; set; }
 
         [JsonProperty("token_period")]
-        public int TokenPeriod { get; set; }
+        public long TokenPeriod { get; set; }
 
         [JsonProperty("token_policies")]
         public List<string> TokenPolicies { get; set; }
 
         [JsonProperty("token_ttl")]
-        public int TokenTimeToLive { get; set; }        
+        public long TokenTimeToLive { get; set; }        
 
         [JsonProperty("token_type")]
         public AuthTokenType TokenType { get; set; }
