+++ src/VaultSharp/V1/AuthMethods/AppRole/Models/FullSecretIdInfo.cs
@@ -25,10 +25,10 @@ namespace VaultSharp.V1.AuthMethods.AppRole.Models
         public string SecretIdAccessor { get; set; }
         
         [JsonProperty("secret_id_num_uses")]
-        public int SecretIdNumberOfUses { get; set; }
+        public long SecretIdNumberOfUses { get; set; }
 
         [JsonProperty("secret_id_ttl")]
-        public int SecretIdTimeToLive { get; set; }
+        public long SecretIdTimeToLive { get; set; }
 
         [JsonProperty("token_bound_cidrs")]
         public List<string> TokenBoundCIDRs { get; set; }
